+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_cli_timeout.py
@@ -29,9 +29,31 @@ options:
       - To prevent CLI sessions from timing out, specify a value of 0 (zero).
     type: int
     required: true
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 
 notes:
   - Only supported with REST and requires ONTAP 9.6 or later.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 """
 
 EXAMPLES = """
@@ -74,8 +96,12 @@ class NetAppOntapCliTimeout:
             state=dict(required=False, type='str', choices=['present'], default='present'),
             timeout=dict(required=True, type='int')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )
         self.na_helper = NetAppModule(self.module)