+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_ems_config.py
@@ -54,10 +54,32 @@ options:
       - Requires ONTAP 9.10 or later.
     type: bool
     required: false
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 
 notes:
   - Only supported with REST and requires ONTAP 9.6 or later.
   - Module is not idempotent when proxy_password is set.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 """
 
 EXAMPLES = """
@@ -110,8 +132,12 @@ class NetAppOntapEmsConfig:
             proxy_password=dict(required=False, type='str', no_log=True),
             pubsub_enabled=dict(required=False, type='bool')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )
         self.uuid = None