+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_ems_filter.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2023-2025, NetApp, Inc
+# (c) 2023-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -16,8 +16,6 @@ version_added: 22.4.0
 author: NetApp Ansible Team (@carchi8py) <ng-ansible-team@netapp.com>
 description:
   - Create, delete, or modify EMS filters on NetApp ONTAP. This module only supports REST.
-notes:
-  - This module only supports REST.
 
 options:
   state:
@@ -69,6 +67,31 @@ options:
           value_pattern:
             description: Value pattern for parameter matching
             type: str
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - This module only supports REST.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 '''
 
 EXAMPLES = """
@@ -158,9 +181,13 @@ class NetAppOntapEMSFilters:
                 ))
             ))
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )
         self.na_helper = NetAppModule(self.module)
@@ -168,7 +195,7 @@ class NetAppOntapEMSFilters:
         self.rest_api = netapp_utils.OntapRestAPI(self.module)
         self.use_rest = self.rest_api.is_rest()
         if not self.use_rest:
-            self.module.fail_json(msg="This module require REST with ONTAP 9.6 or higher")
+            self.module.fail_json(msg="This module requires REST with ONTAP 9.6 or higher")
 
     def get_ems_filter(self):
         api = 'support/ems/filters'