+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_vserver_audit.py
@@ -146,11 +146,33 @@ options:
                   - Specifies the weekdays schedule to rotate audit log. Specify -1 to rotate the audit logs every day.
                 type: list
                 elements: int
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 
 notes:
   - This module supports REST only.
   - At least one event should be enabled.
   - No other fields can be specified when enabled is specified for modify.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 """
 
 EXAMPLES = """
@@ -324,9 +346,13 @@ class NetAppONTAPVserverAudit:
                 user_account=dict(type='bool'),
             ))
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )